Upgrade to import-spec 1.0.0-rc02
In this release, there are no built-in sources so they must be
provided on the Dataflow side, together with their custom
validation constraints.

This upgrade somehow exhibited a conflict between Jackson and
SnakeYaml. Downgrading the latter turns out to be the much
easier option, as Beam embed Jackson which causes a lot of issues
when trying to align with that.